refactor: 이동 경로 메서드 전략 패턴 추출
