+++ src/main/java/janggi/domain/movestrategy/ChoSoldierStrategy.java
@@ -1,6 +1,7 @@
 package janggi.domain.movestrategy;
 
 import janggi.domain.board.Position;
+import java.util.List;
 
 public class ChoSoldierStrategy implements MoveStrategy {
 
@@ -16,5 +17,10 @@ public class ChoSoldierStrategy implements MoveStrategy {
             return true;
         }
         return (Math.abs(nextX - preX) == 1) && (nextY == preY);
+    }
+
+    @Override
+    public List<Position> findPath(Position from, Position to) {
+        return List.of(to);
     }
 }
