+++ src/main/java/janggi/domain/movestrategy/GeneralStrategy.java
@@ -3,8 +3,9 @@ package janggi.domain.movestrategy;
 import static java.lang.Math.abs;
 
 import janggi.domain.board.Position;
+import java.util.List;
 
-public class GeneralStrategy implements MoveStrategy{
+public class GeneralStrategy implements MoveStrategy {
 
     @Override
     public boolean canMoveByBasicMovingRule(Position from, Position to) {
@@ -22,5 +23,10 @@ public class GeneralStrategy implements MoveStrategy{
         }
 
         return abs(preX - nextX) + abs(preY - nextY) <= 1;
+    }
+
+    @Override
+    public List<Position> findPath(Position from, Position to) {
+        return List.of(to);
     }
 }
