+++ src/main/java/janggi/domain/movestrategy/GuardStrategy.java
@@ -3,6 +3,7 @@ package janggi.domain.movestrategy;
 import static java.lang.Math.abs;
 
 import janggi.domain.board.Position;
+import java.util.List;
 
 public class GuardStrategy implements MoveStrategy {
 
@@ -22,5 +23,10 @@ public class GuardStrategy implements MoveStrategy {
         }
 
         return abs(preX - nextX) + abs(preY - nextY) <= 1;
+    }
+
+    @Override
+    public List<Position> findPath(Position from, Position to) {
+        return List.of(to);
     }
 }
