+++ src/main/java/janggi/domain/movestrategy/MoveStrategy.java
@@ -1,8 +1,11 @@
 package janggi.domain.movestrategy;
 
 import janggi.domain.board.Position;
+import java.util.List;
 
 public interface MoveStrategy {
     boolean canMoveByBasicMovingRule(Position from, Position to);
+
+    List<Position> findPath(Position from, Position to);
 }
