+++ src/main/java/janggi/domain/piece/ChariotPiece.java
@@ -12,24 +12,6 @@ public class ChariotPiece extends Piece {
     }
 
     @Override
-    public List<Position> findPath(Position from, Position to) {
-        List<Position> path = new ArrayList<>();
-
-        int currentX = from.getX();
-        int currentY = from.getY();
-
-        int stepX = Integer.compare(to.getX(), from.getX());
-        int stepY = Integer.compare(to.getY(), from.getY());
-
-        while (currentX != to.getX() || currentY != to.getY()) {
-            currentX += stepX;
-            currentY += stepY;
-            path.add(new Position(currentX, currentY));
-        }
-        return path;
-    }
-
-    @Override
     public boolean canMoveBySpecialMovingRule(Map<Position, Piece> positionPieces, Position to) {
         if (positionPieces.size() >= 2) {
             return false;
