+++ src/main/java/janggi/domain/piece/ElephantPiece.java
@@ -13,60 +13,6 @@ public class ElephantPiece extends Piece {
     }
 
     @Override
-    public List<Position> findPath(Position from, Position to) {
-        List<Position> path = new ArrayList<>();
-
-        int preX = from.getX();
-        int preY = from.getY();
-
-        int nextX = to.getX();
-        int nextY = to.getY();
-
-        if (Math.abs(preX - nextX) == 3) {
-            if (nextX > preX) {
-                path.add(new Position(preX + 1, preY));
-                if (nextY > preY) {
-                    path.add(new Position(preX + 2, preY + 1));
-                    path.add(new Position(preX + 3, preY + 2));
-                    return path;
-                }
-                path.add(new Position(preX + 2, preY - 1));
-                path.add(new Position(preX + 3, preY - 2));
-                return path;
-            }
-            path.add(new Position(preX - 1, preY));
-            if (nextY > preY) {
-                path.add(new Position(preX - 2, preY + 1));
-                path.add(new Position(preX - 3, preY + 2));
-                return path;
-            }
-            path.add(new Position(preX - 2, preY - 1));
-            path.add(new Position(preX - 3, preY - 2));
-            return path;
-        }
-        if (nextY > preY) {
-            path.add(new Position(preX, preY + 1));
-            if (nextX > preX) {
-                path.add(new Position(preX + 1, preY + 2));
-                path.add(new Position(preX + 2, preY + 3));
-                return path;
-            }
-            path.add(new Position(preX - 1, preY + 2));
-            path.add(new Position(preX - 2, preY + 3));
-            return path;
-        }
-        path.add(new Position(preX, preY - 1));
-        if (nextX > preX) {
-            path.add(new Position(preX + 1, preY - 2));
-            path.add(new Position(preX + 2, preY - 3));
-            return path;
-        }
-        path.add(new Position(preX - 1, preY - 2));
-        path.add(new Position(preX - 2, preY - 3));
-        return path;
-    }
-
-    @Override
     public boolean canMoveBySpecialMovingRule(Map<Position, Piece> positionPieces, Position to) {
         if (positionPieces.size() >= 2) {
             return false;
