+++ src/main/java/janggi/domain/piece/SoldierPiece.java
@@ -11,11 +11,6 @@ public class SoldierPiece extends Piece {
     }
 
     @Override
-    public List<Position> findPath(Position from, Position to) {
-        return List.of(to);
-    }
-
-    @Override
     public boolean canMoveBySpecialMovingRule(Map<Position, Piece> positionPieces, Position to) {
 
         for (Piece piece : positionPieces.values()) {
